fix setting form validation in PR
