+++ src/pages/dar_application/DataAccessRequestApplication.jsx
@@ -21,10 +21,7 @@ import { Storage } from '../../libs/storage';
 import { assign, cloneDeep, get, head, isEmpty, isNil, isString, keys, map } from 'lodash/fp';
 import './DataAccessRequestApplication.css';
 
-import {
-  getNewFormValidation,
-  validateDARFormData
-} from '../../utils/darFormUtils';
+import {validateDARFormData} from '../../utils/darFormUtils';
 import DucAddendum from './DucAddendum';
 import UsgOmbText from '../../components/UsgOmbText';
 import {DAAUtils} from '../../utils/DAAUtils';
@@ -37,6 +34,7 @@ import {ConditionalAccordion} from "../../components/forms/ConditionalAccordion.
 import {ProgressReportApplication} from "./ProgressReportApplication";
 import {ScrollableTabs} from "./ScrollableTabs";
 import {validationFailed} from "../../utils/darFormUtils.js";
+import {isArray, set} from 'lodash';
 
 // Constants
 const RESEARCHER_INFO_TAB_ID = 'researcher-info';
@@ -163,7 +161,13 @@ const DataAccessRequestApplication = (props) => {
 
   const formValidationChange = useCallback((section, { key, validation }) => {
     setFormValidation((formValidation) => {
-      getNewFormValidation(formValidation, section, key, validation);
+      const newFormValidation = cloneDeep(formValidation);
+      if (isArray(key)) {
+        set(newFormValidation, [section, ...key], validation);
+      } else {
+        set(newFormValidation, [section, key], validation);
+      }
+      return newFormValidation;
     });
   }, []);
